epost info uses int64

+++ internal/pkg/block/epost_info.go
@@ -23,16 +23,16 @@ type EPoStInfo struct {
 type EPoStCandidate struct {
 	_                    struct{} `cbor:",toarray"`
 	PartialTicket        []byte
-	SectorID             types.Uint64
-	SectorChallengeIndex types.Uint64
+	SectorID             uint64
+	SectorChallengeIndex uint64
 }
 
 // NewEPoStCandidate constructs an epost candidate from data
 func NewEPoStCandidate(sID uint64, pt []byte, sci uint64) EPoStCandidate {
 	return EPoStCandidate{
-		SectorID:             types.Uint64(sID),
+		SectorID:             sID,
 		PartialTicket:        pt,
-		SectorChallengeIndex: types.Uint64(sci),
+		SectorChallengeIndex: sci,
 	}
 }
 
@@ -72,8 +72,8 @@ func ToFFICandidates(candidates ...EPoStCandidate) []ffi.Candidate {
 func FromFFICandidate(candidate ffi.Candidate) EPoStCandidate {
 	return EPoStCandidate{
 		PartialTicket:        candidate.PartialTicket[:],
-		SectorID:             types.Uint64(candidate.SectorID),
-		SectorChallengeIndex: types.Uint64(candidate.SectorChallengeIndex),
+		SectorID:             candidate.SectorID,
+		SectorChallengeIndex: candidate.SectorChallengeIndex,
 	}
 }
